QQmlListAccessor: Store integer if m_type is Integer
When we pass a QVariant to QQmlListAccessor::setList which contains
something convertible to a number, we set the type to Integer. Later, we
simply return whatever is stored internally in the variant, and cast it
into an int. That goes horribly wrong if the variant actually contains a
floating point number.
To fix this, we now store the actual integer in the variant after the
checks are done.

Change-Id: I3aa173a1558e95361bb4b19e97542d8a435a6dd7
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 42afb8249cfcf8cec66aac9e55d13ac508699ec8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>