Selection support: add private interface QQuickSelectable
Add a private interface that can be used for controlling selections.
This interface will initially be implemented by TableView, but
also ListView, GridView, and potentially any component can
implement this interface.

Next, a visual selection tool called SelectionRectangle will be
added to Controls. This tool has e.g styled selection handles, and
does therefore not belong in Quick. This tool will not communicate
directly with TableView, but through the added interface. That
way, the tool can be reused for any component / control in Quick or
Controls that needs a selection tool, and implements the interface.

Change-Id: I1cf9d4ba9c063676ef6458e0cfd9d077bed0d4d0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>