Add new control: SelectionRectangle
This patch will add a SelectionRectangle to controls.

A SelectionRectangle can be used together with TableView
to let the user select cells using either pointer Drag
or PressAndHold. It lets the developer set their own
handle delegates, or simply use the included ones added
to the Basic style.

[ChangeLog][Controls][SelectionRectangle] A SelectionRectangle
has been added to Controls. A SelectionRectangle can be used
to make selections in TableView.

Fixes: QTBUG-74750
Change-Id: Ia7af578c1c905b0e92df193d7820ad6ef2c9b697
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>