Revert "AbstractButton: set automatically as checkable when being checked"
This reverts commit 8ba12ed7f25d8cfb7619b6eb0a26fe4ae3e823b1.

The referenced commit was introduced to simplify buttons setup:

    Button { checked: true }
    Button { checkable: true }

With the commit the first button becomes 'checkable' automatically.

However, the implemented behavior violates the principle of least
surprise, listed in "API Design Principles". The following checkbox
will be checkable despite of the explicitly set properties:

    CheckBox {
        checked: true
        checkable: false
    }

This is a "surprising" behavior which makes some rare and already
complex use-cases unnecessarily complicated.

The save of explicit "checkable: true" does not justify the surprice
and the ugliness of the code needed to overcome the forced checkable
property if needed.

[ChangeLog][Important Behavior Changes] Setting the AbstractButton's
'checked' property to 'true' does not automatically set its 'checkable'
to true anymore.

Fixes: QTBUG-93807
Change-Id: Ia35138ee6a09bb9f361f52a84e53578b0b6e57e3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>