Require intended types for setting up AOT lookups
The object to be used for the lookup can be different from what the
caller expects. We need to verify that the caller talks about the same
type as the object. If it doesn't, throw an exception.

For the same reason, also throw an exception rather than asserting if
the property caches don't match.

Change-Id: I3b45691bfe2204820d878a624814d412d1358c24
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>