QML_DECLARE_TYPE: Do register type with normalized name
The normalized name of a pointer does not contain a space.
QML_DECLARE_TYPE should be superfluous in general for any type
registered to QML, but as it already exist, we should at least ensure
that it won't cause additional work.

Note that this will not gain us much until QTBUG-93841 is fixed in
qtbase.

Task-number: QTBUG-93841
Change-Id: Ibf02968e8ee3205afc4dd7531e33aaad070874a5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>