TapHandler: don't emit tapped() if non-filtering grabber on release
This comes up for example in qtquick3d/examples/quick3d/dynamictexture:
A TapHandler detects clicks on a 3D door model, to trigger the animation
to open the door. But onto the door is mapped a subscene containing a
TextEdit.  If the user clicks in the text to set the cursor position:
1) TapHandler gets the press first (it's a handler, handlers go before
   items like the View3D), and takes a passive grab.
2) View3D does picking, discovers the subscene on the door surface and
   begins subscene delivery
3) TextEdit takes the exclusive grab
4) On release, the main DA goes first again; TapHandler gets the event
   because of its passive grab; but in QQTapHandler::handleEventPoint()
   we can see that there is another exclusive grabber, which indicates
   that it (the TextEdit) has taken full responsibility for handling the
   event sequence. TapHandler cancels the press and doesn't emit tapped().
5) TextEdit gets the event via the subscene agent because of its exclusive
   grab; it sees a click and sets cursor position.

On the other hand, if we did not check QQuickItem::filtersChildMouseEvents(),
we would break some Flickable "interop" tests which expect that handlers
can react in spite of Flickable having a grab.  Being excessively
"grabby" seems to go with the child-filtering concept, so let's continue
to let TapHandler react in spite of the grabber in that case.

It remains to be seen if we need alternate behavior in other scenarios,
which would indicate the need for another enum property or so.

Fixes: QTBUG-94012
Task-number: QTBUG-92944
Change-Id: I7d2b03fcb0d286d581547a14e71c27086a9b25f6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>