qmllint: Speed up looking for builtins
Previously we kept searching for more builtins qmltypes when we've already found them.
This wouldn't be a big issue except for the fact that if you have a directory with lots of subdiretories in your import path this might take a long time.
Now we abort as soon as we've seen each file we've been looking for once.

Change-Id: I4d7ed74840366dab55dd9dfc4f7c939a0070bdc9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>