Deprecate QQuickDragHandler::translation()
Amends 60d11e1f69470d588666b76092cd40ae5644a855, in which we deprecated
the QML property.  In C++ this is still a private class, so most non-QML
user code won't be using it; but we should eventually ensure that it is
no longer in use in our code too (it's nice to have this short name, but
handlers are inconsistent about which of the short-name properties are
persistent and which are gesture-scoped).

Task-number: QTBUG-94168
Change-Id: Ic2bac67c4c5758e7edce8834eae7eb74dc3e2471
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>