QQuickItem change contains function behavior
QQuickItem::contains() and QQuickItem::childAt() were different:
QQuickItem::contains() included edges, but QQuickItem::childAt()
did not. Because of this, it is impossible to replace the condition
at QQuickItem::childAt() function whether the point belongs to an item,
to function QQuickItem::contains() (this is necessary to start using
containment mask), without breaking the childAt function, because it
would start to consider the edges.
This change prohibits delivery of events to zero-size items, and fixes
the case when the same point could otherwise be inside multiple items:
for example (0,0,10,10), (10,10,20,20) and the point (10,10).

[ChangeLog][Important Behavior Changes][QQuickItem] Item.contains() now
checks for coordinates less than the width and height, for consistency
with childAt(), and because the item is drawn from (0,0) to
(width-1, height-1).

Fixes: QTBUG-55958
Change-Id: I1556cc14038cef4d2a6bb1ab7d28aa449cfd7a25
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 37540467ca516c3732bd276f49e27db824b4ea0b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>