Allow registration of enums from related types to be switched off
This was a terrible misfeature. It registers random enums from all over
the place into a type. See the test for an example.

[ChangeLog][QtQml] You can now suppress the registration of enums of
"related" C++ types in a QML type. Such registration is often unintended
and rather unpredictable. Set the "RegisterEnumsFromRelatedTypes"
Q_CLASSINFO to "false" in order to suppress the registration.

Task-number: QTBUG-83703
Change-Id: Ic9046daa7a3c833ce65396a8731dfa28d269317c
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>