qmlcompiler: Process properties only once parsing is completed
Now we do most properties after the fact which eliminates false positives on unknown types that are not known at the time of traversing the AST. It also allows for chaining inline components.
This effectively introduces a two pass system but there are some exceptions (i.e. signals) where some deduction that should be run after the fact still runs during parsing.

Fixes: QTBUG-93652
Pick-to: 6.2
Change-Id: Ic1ac0e8ce2d5d5dfbe80c16a341b10cf2b078d81
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>