QQuickComboBox: fix acceptableInput being wrong if no validator was set
This patch fixes an issue with hasAcceptableInput(), if the property
would be read before the contentItem had been set by the qml engine.
This would cause hasAcceptableInput to return false by default, even
though the default value is supposed to be true, if no validators or
inputMasks are being used.

The solution that I've chosen, is to give the QQuickComboBox its own
acceptableInput variable, and connect the contentItem's
acceptableInputChanged() signal to a function that polls for the
contentItem property, and updates its variable accordingly.

Fixes: QTBUG-94307
Pick-to: 6.2 6.1 5.15
Change-Id: I587d76162e75544a7ed1df9e3b9104bd73013bb0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>