Use Inactive color group in inactive windows
A palette has three color groups (Active, Inactive, Disabled), but we
only listened to changes of the enabled state, and always used the
Active color group for an enabled item.

In order for the Inactive color group to take effect, we need to also
trigger the update when the "active" property changes. Instead of
connecting yet another signal, deliver WindowActivate/Deactivate to
as well, and forward from contentItem to all children which can then
update the current color group.

Add a test case.

Fixes: QTBUG-93752
Pick-to: 6.2
Change-Id: I4f2b6121e822115aaa5c4faaa5d402932dacc67b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>