Don't deliver QDragEnterEvent to a drop area that is a child of the dragged item
If the drag.target of the mouse area that starts the drag is a parent of
a drop area, then that drop area cannot be dropped on (as it's moving
with the dragged item).

Due to QQuickDeliveryAgentPrivate::deliverDragEvent calling itself
recursively, this is the earliest place where we can block this event
from being delivered.

Fixes: QTBUG-64128
Change-Id: I495492161d67f7ac60eafb99982fb01ec2374c50
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit e36b5ce4df4239fdf7c54afcc9c261fe1cceafc1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>