Don't deliver QDragEnterEvent to a drop area that is a child of the dragged item
If the drag.target of the mouse area that starts the drag is a parent of
a drop area, then that drop area cannot be dropped on (as it's moving
with the dragged item).

Due to QQuickDeliveryAgentPrivate::deliverDragEvent calling itself
recursively, this is the earliest place where we can block this event
from being delivered.

Fixes: QTBUG-64128
Pick-to: 6.2
Change-Id: I495492161d67f7ac60eafb99982fb01ec2374c50
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>