Make tst_qquickshadereffect::testConnection() more robust
The test is there to verify that the sourceChanged property
notifier gets a connection, but it also counted all other
connections. Since b65159a5ea8db05165b2eaab8e180a12f30063e4 in
qtbase the parentChanged, windowChanged and enabledChanged
signals are also connected to something so this broke the
test.

Fix is to make the test explicitly look for the sourceChanged
signal and ignore all others.

Pick-to: 6.2
Change-Id: Ia188384b37c9c078e78d09670bd955a69d10c7de
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>