Install .qml files from source dir and honor their resource alias
For targets that are built as part of Qt, we currently require all .qml
files to be passed to qt_internal_add_qml_module(). We can use that to
iterate over the set of qml files and access their source file
properties at configure time. The previous approach tried to use the
list obtained at generate time, but the source file properties (and
therefore their resource aliases) were not accessible and the install
locations assumed no resource path or alias.

By installing from the source dir, we also address the dependency issue
where an install wouldn't cause the build dir copies to be updated. The
install target will always install the latest source files with this
change.

Fixes: QTBUG-94519
Fixes: QTBUG-94520
Pick-to: 6.2
Change-Id: I6a97a9631ff77c3f7696899a347b83c2eb4b0ca4
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>