Ensure model is in context if required properties are not used
We need to set the set the contextObject already in
QQmlDelegateModelPrivate::object, as the context property might be
accessed in a child component before we are able to run
setRequiredProperties (which would set-up the context in case there are
no required properties).
Instead of delaying setting the context object, we clear it now in
setRequiredProperties.
This has the drawback that one might be able to access context
properties on initial component setup which should not exist
(due to the presence of required properties).
This could be avoided by inspecting the Compilation Unit and only
setting the context if we do not find required properties, however that
needs further work to expose the necessary information.

Fixes: QTBUG-94223
Change-Id: I678de81408539417fc650f94da4c9f0d3167653d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 58cd06033cacadab541efaa16a3eecec37dab0fa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>