ToolTip: test that smaller attached tooltips don't shrink larger ones
This ensures that any solution to QTBUG-83630 does not cause a regression.

The test that this patch adds fails for the Imagine style ToolTip (which
is tracked by QTBUG-101704), but that can be fixed separately, as there
is value in ensuring that we don't regress again via the other passing
tests.

Task-number: QTBUG-83630
Task-number: QTBUG-94764
Task-number: QTBUG-101704
Change-Id: I3a827b56c0aac9d402a71dc5c4124356a270b4ba
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 0b9a53598b9485afc84cf67271370d77794d6b2d)
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>