CMake: Add test to check backing lib resources are linked
Add a test to check that a Qml backing lib's resource object files are
linked into the final executable across a static qml plugin.

This didn't work before because plugin targets link privately
against their backing lib, and that caused the resource object files
not to be propagated to the final executable using target_sources
linking approach.

It should not be a problem anymore because the minimum CMake version
for using static Qt has been raised.

Usage of qmlshapesplugin in an application exhibited this issue.

The test uses QtQuickShapes to instantiate a ShapePath. It then
installs a message handler to intercept a warning in case if one of
the shaders embedded in the resource object files is missing.

Pick-to: 6.2
Task-number: QTBUG-92933
Change-Id: Ib4be313201f705c3daad5f48dc92360cac107ddb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>