QtQml: Move locateLocalQmldir and make it a template
It conceptually belongs into QQmlImportDatabase because it messes with
the qmldir cache.

This way, if we reject a qmldir file for whatever reason, we can still
check for further matching qmldir files in other places. We might, for
example determine that we need to load a plugin, but we have been given
a qmldir file in the resource file system. In that case, we want to
check for other instances of the same module in the host file system.

Change-Id: I8fe4a0f188f3732b9d10d017a94562e8bd1fb242
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit dbca762f45d3249974652cccb5d6421440f852ae)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>