QQuickItem::{width,height}Valid: correctly handle binding being undefined
The width/height is valid if we have a binding, unless the binding'
value is actually undefined. In that case, the width/height has been
reset. We still need to keep the binding, as it might change to a
defined state. Thus we need to explicitly check in the undefined case
whether the binding might be undefined.

Pick-to: 6.2
Fixes: QTBUG-94703
Change-Id: I0fd4ac608fdf4eeab09bb2c092045ef803233adc
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>