Allow multiple imports with the same URI for the same Blob
This can easily happen if the same module is imported several times with
different prefixes, or if there are multiple implicit imports from other
modules. Before, the last request for a qmldir URL overwrote any
previous requests, leading to the previous requests never getting
notified about their completion.

The QQmlTypeLoader test exercises this code in a particularly mean way
by intercepting local URLs and making them remote. This, combined with
the fact that we now can have multiple local locations for the same
module, made it much more likely to hit the above condition.

This change makes the test even meaner by randomizing the response time.
This way we increase the probability that some qmldir request is posted
twice before a response is received.

Pick-to: 6.2
Change-Id: Ie8b802cb65cfb78b0cde22f485ad7a31f60ee5bf
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>