QQDeliveryAgent: keep lastMousePosition updated during dragging
Given that we store lastMousePosition (which I hope we can move to
QPointingDevice eventually), the idea is it should always be correct; so
don't bypass updating it during a mouse move just because there's a
grabber.  It resulted in synthetic hover events being sent after
release, with an outdated last-known mouse position.

Fixes: QTBUG-92563
Pick-to: 6.2 6.1
Change-Id: Ic708bfcbdfcd0a9f6dc2cfd42f0671b0661e76bb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>