QJSValue: Avoid BIC break
The toScriptValue function template called the exported create function.
Thus, we can not change the signature of create, even though it was
private. To avoid the BIC break, we keep the old version (and ifdef it
so that it will go away in Qt 7).

Task-number: QTBUG-94407
Change-Id: I5b07f978dca156f52bdb529d3e15aea8c0c3c97e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit cd8e839e032f31a7421f33fd65dc3eaba4d84052)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>