CMake: Ensure ${target}_qmllint targets depend on qmltypes files
The *.qmltypes files are usually generated automatically. They are
needed by qmllint when it is run when bringing the ${target}_qmllint
target up to date. If the ${target}_qmllint target is built before the
${target} itself, we need to ensure the ${target}.qmltypes file has
been generated. Furthermore, ${target} might depend on other QML
modules and the *.qmltypes files for those QML modules also need to
exist at the time qmllint runs. The only way we can robustly ensure
this is to make each ${target}_qmllint target depend on the
generation of all *.qmltypes files, since we don't have a way to
know the QML module dependencies at configure time (at least, not
for shared builds where qmlimportscanner isn't used).
This commit adds the required dependencies to enforce this
conservative constraint.

Task-number: QTBUG-94761
Pick-to: 6.2
Change-Id: I13561517db3d268dc3d2d6ed8e183516d5b57dc4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>