QQuickOverlay: Handle hover events
With recent changes to hover handling logic, hover events would have
been sent to the "main window", even if a popup were modal. As this is
unwanted (a modally blocked button should not auto-raise when hovered),
we now prevent delivery of those events in QQuickOverlay and
overlayEvent.
The logic is the same as for mouse and touch events: If we get a
matching event, we ask the popup to handle it. The popup then uses the
existing blockInput logic to decide whether the event should be blocked
or forwarded.

Change-Id: I2194fd8e832592efd5b7b9697412bdaeaea74b83
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
(cherry picked from commit 48da34f3d797f58b3eda9ed0a98aa1776d75d946)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>