Propagate LocaleChange events down the item tree
We do this already for LanguageChange, and items need to get informed
about a LocaleChange event as well, esp since not all items will have
their locale updated via QQuickApplicationWindow and QQuickControl's
recursive update. This way, items that are not QQuickControls such as
Label or TextField can respond to this change as well.

Fixes: QTBUG-75239
Pick-to: 6.2
Change-Id: I5d7309a83f67ccfff550e41bbc1f8bdb04abf3a5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>