Resolve data race in QQmlPropertyData in a minimally invasive way
We keep all the idiosyncrasies of when what flag or member is set. We
only make sure not to call resolve() concurrently by moving the propType
into an atomic int which also replaces the "fully resolved" flag.

We need to move another 16bit member in there as only 32bit atomics are
guaranteed to work everywhere and we want to keep the fixed size of the
class. So we choose the relative property index and only set it when we
are sure there isn't a data race. This is fine because for missing
relative property indices there is a fallback.

Task-number: QTBUG-93973
Change-Id: Ia357444cad85fe2df234e76226dd8adc10fe8549
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>