Introduce Q_IMPORT_QML_PLUGIN
In Qt 6.2, Q_IMPORT_QML_PLUGIN behaves exactly the same as
Q_IMPORT_PLUGIN. The goal is however to make it obsolete in the future,
by taking care of the plugin handling purely in CMake. The idea is to
have Q_IMPORT_QML_PLUGIN expand to nothing in a later release; having
the additional layer indirection via Q_IMPORT_QML_PLUGIN allows that
change, which would be impossible if we were to directly use
Q_IMPORT_PLUGIN. Having the macro later expand to nothing is required
because otherwise we might attempt to instantiate a static plugin
instance with the same name twice, leading to duplicated symbol errors.

Pick-to: 6.2
Task-number: QTBUG-95148
Change-Id: I29bf4637f37d6358f07935c476544888c1c4c023
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>