Fix qmldir cache inconsistency
If there are multiple qmldir entries for the same QML module in the
import list, the one that comes first should take precedence. The same
applies when constructing the qmldir cache.

We can achieve that by inserting the entries into the cache in the same
order they appear in the completeQmldirPaths list.

Pick-to: 6.2
Change-Id: Ic56a0952b16c4be016180c695d79089fc132319a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>